Create gitutils module
